feat: add connect to spike service with grpc
