+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '2.1'
 services:
   user-service:
-    image: user-service:local
+    image: user-service:latest
     ports:
       - '8086:8080'
     environment:
